refactor: move KnownAction to types/action.ts
